% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pal16}
\alias{pal16}
\title{Pal 16}
\format{An object of class \code{character} of length 16.}
\usage{
pal16
}
\description{
Palette of colors to use for 16 classes or less
}
\keyword{datasets}
