% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pal21}
\alias{pal21}
\title{Pal 21}
\format{An object of class \code{character} of length 22.}
\usage{
pal21
}
\description{
Palette of colors to use for 21 classes or less
}
\keyword{datasets}
