% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot feature on low-dimensional space}
\usage{
plotFeature(object, feature, dims = c(1, 2), reduction = "umap",
  group = NULL, type = "data", qclip = 0.99, label = TRUE,
  alpha = 0.7, size = 1)
}
\arguments{
\item{object}{A Seurat object}

\item{feature}{Feature to be plotted (gene expression or metadata)}

\item{dims}{Dimensions to plot. Only two are valid.}

\item{reduction}{Dimensionality reduction (e.g. \code{pca}, \code{umap}, \code{tsne}, ...)}

\item{group}{Group variable in metadata}

\item{type}{If gene expression is used, specify, rata type: \code{counts} or \code{data} slots}

\item{qclip}{Quantile value to clip gene expression values or continuous variables. This parameter reduces the effect of
outlier cells with high gene expression according to a quantile of the distribution (default 0.99). All
cells for each gene expressing a value greater to the quantile value in the distribution are rescaled to
the corresponding value of that quantile}

\item{label}{Label clusters when a categorical variable is used. Removes legend.}

\item{alpha}{Alpha level to adjust transparency of colors}

\item{size}{Size of points}
}
\value{
A plot with feature of interest
}
\description{
This function allows to plot the values of a feature (gene or metadata column) on
a low-dimensional space
}
\examples{

# Plot gene expression of CD8A and CD3D from log-normalized data over a UMAP plot

plotFeature(pbmc, "seurat_clusters")

}
\author{
José Alquicira Hernández
}
