% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{plotMerge}
\alias{plotMerge}
\title{Plot expression of genes and their overlap}
\usage{
plotMerge(object, gene1, gene2, dims = c(1, 2), reduction = "umap",
  group = NULL, type = "data", qclip = 0.99, alpha = 0.7,
  size = 1, bgColor = "#171716", returnGrid = TRUE)
}
\arguments{
\item{object}{A Seurat object}

\item{gene1}{Gene 1}

\item{gene2}{Gene 2}

\item{dims}{Dimensions to plot. Only two are valid.}

\item{reduction}{Dimensionality reduction (e.g. \code{pca}, \code{umap}, \code{tsne}, ...)}

\item{group}{Group variable in metadata}

\item{type}{Data type: \code{counts} or \code{data} slots}

\item{qclip}{Quantile value to clip gene expression values. This parameter reduces the effect of
outlier cells with high gene expression according to a quantile of the distribution (default 0.99). All
cells for each gene expressing a value greater to the quantile value in the distribution are rescaled to
the corresponding value of that quantile}

\item{alpha}{Alpha level to adjust transparency of colors}

\item{size}{Size of points}

\item{bgColor}{Background color}

\item{returnGrid}{If \code{TRUE}, a grid is generated using cowplot. Otherwise, a list with the plots is
returned}
}
\value{
A grid plot with individual genes and merge gene expression values
}
\description{
This function allows to plot the expression of genes in individual color scales
and their corresponding "merge" visualization
}
\examples{

# Plot gene expression of CD8A and CD3D from log-normalized data over a UMAP plot

plotMerge(pbmc, "CD8A", "CD3D", reduction = "umap", type = "data")

}
\author{
José Alquicira Hernández
}
